+++ modules/datastore/modules/datastore_mysql_import/tests/src/Kernel/Storage/MySqlDatabaseTableTest.php
@@ -3,6 +3,7 @@
 namespace Drupal\Tests\datastore_mysql_import\Kernel\Storage;
 
 use Drupal\common\DataResource;
+use Drupal\common\Storage\ImportedItemInterface;
 use Drupal\datastore_mysql_import\Factory\MysqlImportFactory;
 use Drupal\datastore_mysql_import\Storage\MySqlDatabaseTable;
 use Drupal\KernelTests\KernelTestBase;
@@ -40,7 +41,6 @@ class MySqlDatabaseTableTest extends KernelTestBase {
     $import_job = $import_factory->getInstance($identifier, ['resource' => $data_resource])
       ->getImporter();
     $this->assertInstanceOf(MySqlDatabaseTable::class, $import_job->getStorage());
-
 
     $result = $import_job->run();
     $this->assertEquals(Result::DONE, $result->getStatus(), $result->getError());
@@ -93,6 +93,43 @@ class MySqlDatabaseTableTest extends KernelTestBase {
     $this->expectException(\Exception::class);
     $this->expectExceptionMessage('Could not instantiate the table due to a lack of schema.');
     $db_table->count();
+  }
+
+  /**
+   * @covers ::hasBeenImported
+   */
+  public function testHasBeenImported() {
+    // Do an import.
+    $identifier = 'my_id';
+    $file_path = dirname(__FILE__, 4) . '/data/columnspaces.csv';
+    $data_resource = new DataResource($file_path, 'text/csv');
+
+    $import_factory = $this->container->get('dkan.datastore.service.factory.import');
+    $this->assertInstanceOf(MysqlImportFactory::class, $import_factory);
+
+    $import_job = $import_factory->getInstance($identifier, ['resource' => $data_resource])
+      ->getImporter();
+
+    $table = $import_job->getStorage();
+    $this->assertInstanceOf(MySqlDatabaseTable::class, $table);
+    $this->assertInstanceOf(ImportedItemInterface::class, $table);
+
+    // Has not been imported yet.
+    $this->assertFalse($table->hasBeenImported());
+
+    // Perform the import.
+    $result = $import_job->run();
+    $this->assertEquals(Result::DONE, $result->getStatus(), $result->getError());
+    $this->assertEquals(2, $import_job->getStorage()->count());
+
+    // Has been imported.
+    $this->assertTrue($table->hasBeenImported());
+
+    // Make a whole new importer.
+    $import_job = $import_job = $import_factory->getInstance($identifier, ['resource' => $data_resource])
+      ->getImporter();
+    // Storage should report that it's been imported.
+    $this->assertTrue($import_job->getStorage()->hasBeenImported());
   }
 
 }
