+++ modules/datastore/tests/src/Kernel/Storage/DatabaseTableTest.php
@@ -3,6 +3,8 @@
 namespace Drupal\Tests\datastore_mysql_import\Kernel\Storage;
 
 use Drupal\common\DataResource;
+use Drupal\common\Storage\ImportedItemInterface;
+use Drupal\datastore\Plugin\QueueWorker\ImportJob;
 use Drupal\datastore\Service\Factory\ImportServiceFactory;
 use Drupal\datastore\Storage\DatabaseTable;
 use Drupal\KernelTests\KernelTestBase;
@@ -27,9 +29,12 @@ class DatabaseTableTest extends KernelTestBase {
   ];
 
   /**
-   * @covers ::hasBeenImported
+   * Ensure that non-mysql-import tables do not implement hasBeenImported().
+   *
+   * We exercise the whole import service factory pattern here to make sure we
+   * get the DatabaseTable object we expect.
    */
-  public function testHasBeenImported() {
+  public function testNoHasBeenImported() {
     // Do an import.
     $identifier = 'my_id';
     $file_path = dirname(__FILE__, 4) . '/data/columnspaces.csv';
@@ -40,26 +45,16 @@ class DatabaseTableTest extends KernelTestBase {
 
     $import_job = $import_factory->getInstance($identifier, ['resource' => $data_resource])
       ->getImporter();
+    $this->assertInstanceOf(ImportJob::class, $import_job);
 
     $table = $import_job->getStorage();
     $this->assertInstanceOf(DatabaseTable::class, $table);
-
-    // Has not been imported yet.
-    $this->assertFalse($table->hasBeenImported());
+    $this->assertNotInstanceOf(ImportedItemInterface::class, $table);
 
     // Perform the import.
     $result = $import_job->run();
     $this->assertEquals(Result::DONE, $result->getStatus(), $result->getError());
     $this->assertEquals(2, $import_job->getStorage()->count());
-
-    // Has been imported.
-    $this->assertTrue($table->hasBeenImported());
-
-    // Make a whole new importer.
-    $import_job = $import_job = $import_factory->getInstance($identifier, ['resource' => $data_resource])
-      ->getImporter();
-    // Storage should report that it's been imported.
-    $this->assertTrue($import_job->getStorage()->hasBeenImported());
   }
 
 }
